Factory pattern implemented. Restructured app to help imports
